🚚 Move the `Poll` object part to the `Channel` namespace 🚚 Move the `Poll` create object art to the `Poll` namespace
